Extract admin game-catalog routes out of submissions.ts
Moves the operator-only published-games shelf routes (list, regate, delete,
slug-backfill, title-backfill) into a new apps/api/src/catalog/admin-game-routes.ts
module, alongside catalog-routes.ts and catalog-search-routes.ts. The
invalidatePublishedGameCaches/isSlugClaimed/confirmSlugClaim closures stay in
submissions.ts (they have other callers in createGame and the connect route)
and are passed in as options.

submissions.ts drops from 4802 to 4628 lines; the new file is 146 lines,
0 comment-prose debt.
